kafka/pyproject.toml
[tool.poetry]
name = "kafka"
version = "0.1.0"
description = ""
authors = ["talhak911 <talhakhan8300@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.2"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pytest = "^8.1.1"
httpx = "^0.27.0"
sqlmodel = "^0.0.18"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
